test: Move Amazon SignV4 tests to package directory
Moves the existing tests for the Amazon external client from a separate `test` subdirectory into the main `amazon` package directory. This ensures they are properly recognized by standard Go test coverage tooling and aligns with idiomatic Go project layout, properly measuring the 96% coverage for the `SignV4` functionality.
