+++ backend/internal/shared/extclient/amazon/sigv4_suite_test.go
@@ -1,4 +1,4 @@
-package test
+package amazon
 
 import (
 	"encoding/json"
@@ -8,8 +8,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	"mi-tech/internal/shared/extclient/amazon"
 )
 
 type TestSuite struct {
@@ -103,7 +101,7 @@ func TestSigV4Suite(t *testing.T) {
 				body = []byte(tt.Request.Body)
 			}
 
-			err = amazon.SignV4(req, body, suite.Config.AccessKeyID, suite.Config.SecretAccessKey, suite.Config.Region, suite.Config.Service, signTime)
+			err = SignV4(req, body, suite.Config.AccessKeyID, suite.Config.SecretAccessKey, suite.Config.Region, suite.Config.Service, signTime)
 			if err != nil {
 				t.Fatalf("SignV4 failed: %v", err)
 			}
